+++ EnhanceQoL/Settings/MainCategory.lua
@@ -9,7 +9,6 @@ local rootCategories = {
 	{ id = "SOCIAL", label = _G["SOCIAL_LABEL"] },
 	{ id = "ECONOMY", label = L["Economy"] or "Economy" },
 	{ id = "SOUND", label = _G["SOUND"] },
-	{ id = "SYSTEM", label = _G["SYSTEM"] },
 	{ id = "PROFILES", label = L["Profiles"] },
 }
 for _, entry in ipairs(rootCategories) do